refactor: finished docker setup
affects: apis, auto-api-tests, auto-client-api, client, common-lib, discord-bot, evmplus-org-server,
server-common, server-jest-config, util-cli

finished merging docker changes and made sure they were updated
updated developer configuration to
better make use of Docker

+++ capwatch-cron/download.sh
@@ -33,11 +33,11 @@ node --no-warnings dist/importCapwatch.js $CAPWATCH_ZIP_PATH
 
 echo "Done with CAPWATCH import"
 
-rm $CAPWATCH_ZIP_PATH
+[ ! -z "$KEEP_CAPWATCH_ZIP" ] && rm $CAPWATCH_ZIP_PATH
 
 echo "Performing Discord updates"
 
-cd /usr/evm-plus/packages/discord
+cd /usr/evm-plus/packages/discord-bot
 
 test -f /run/secrets/discord_client_token && node --no-warnings dist/index.js updateservers
